+++ CHANGELOG.md
@@ -14,6 +14,7 @@ You can also check the
 - Features
   - Table sorting is not reset anymore on data / filter change
   - Visualize now supports multiple cube contact points
+  - Order of groups when sorting by total segment size is now more intuitive
 - Fixes
   - Hiding a temporal column and then enabling an interactive filter doesn't
     result in broken state anymore